requirements.txt
pytorch_toolbelt==0.4.0
albumentations==0.4.1
opencv_python_headless==4.1.1.26
tqdm==4.41.1
torch==1.5.1
matplotlib==3.1.2
numpy==1.18.0
segmentation_models_pytorch==0.1.2
Shapely==1.6.4.post2
Pillow==8.0.1
